# openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/template:
    post:
      operationId: TemplateController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateDto'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTemplateResponseDto'
        '400':
          description: Validation error
        '409':
          description: Template name already exists
      summary: Create a new template with initial content
      tags: &ref_0
        - Templates
  /api/v1/template/{id}:
    get:
      operationId: TemplateController_getTemplate
      parameters:
        - name: id
          required: true
          in: path
          description: Template UUID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Template retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateWithVersionResponseDto'
        '404':
          description: Template not found
      summary: Retrieve a template with its latest version
      tags: *ref_0
    put:
      operationId: TemplateController_updateTemplate
      parameters:
        - name: id
          required: true
          in: path
          description: Template UUID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateDto'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateWithVersionResponseDto'
        '400':
          description: Validation error
        '404':
          description: Template not found
      summary: Update template content (creates a new version)
      tags: *ref_0
  /api/v1/template/{id}/versions:
    get:
      operationId: TemplateController_getVersions
      parameters:
        - name: id
          required: true
          in: path
          description: Template UUID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Versions retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateVersionResponseDto'
        '404':
          description: Template not found
      summary: Retrieve all versions of a template
      tags: *ref_0
  /api/v1/template/{id}/versions/{versionNumber}:
    get:
      operationId: TemplateController_getVersion
      parameters:
        - name: id
          required: true
          in: path
          description: Template UUID
          schema:
            format: uuid
            type: string
        - name: versionNumber
          required: true
          in: path
          description: Version number
          schema:
            type: number
      responses:
        '200':
          description: Version retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateVersionResponseDto'
        '404':
          description: Template or version not found
      summary: Retrieve a specific version of a template
      tags: *ref_0
  /api/v1/prompt:
    post:
      operationId: PromptController_generate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePromptDto'
      responses:
        '200':
          description: Prompt generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptResponseDto'
        '400':
          description: Validation error or missing required variables
        '404':
          description: Template not found
      summary: Generate a new prompt from a template
      tags:
        - Prompts
info:
  title: Prompt Versioning API
  description: REST API for managing and versioning prompt templates
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateTemplateVariableDto:
      type: object
      properties:
        name:
          type: string
          description: Variable name (must be unique within template)
          example: name
        defaultValue:
          type: object
          description: Optional default value
          example: Customer
          nullable: true
      required:
        - name
    CreateTemplateDto:
      type: object
      properties:
        name:
          type: string
          description: Template name (must be unique, max 50 characters)
          example: Customer Email
          maxLength: 50
        tags:
          description: Array of tag labels for categorization
          example:
            - email
            - customer
          minItems: 1
          type: array
          items:
            type: array
        content:
          type: string
          description: Template content with optional variable placeholders
          example: Hello {{name}}, thank you for your business!
        variables:
          description: Array of variables used in the template
          type: array
          items:
            $ref: '#/components/schemas/CreateTemplateVariableDto'
      required:
        - name
        - tags
        - content
    CreateTemplateResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the template
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Template name
          maxLength: 50
          example: Customer Email
        tags:
          description: Array of tag labels
          example: &ref_1
            - email
            - customer
          type: array
          items:
            type: array
        currentVersion:
          type: number
          description: Version number of the latest version
          minimum: 1
          example: 1
        createdAt:
          type: string
          description: Timestamp when template was created (ISO 8601)
          format: date-time
          example: '2026-05-07T10:30:00Z'
      required:
        - id
        - name
        - tags
        - currentVersion
        - createdAt
    VariableResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the variable
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          description: Variable name used for substitution in template content
          example: firstName
        defaultValue:
          type: object
          description: Optional default value used if not provided at prompt generation time
          nullable: true
          example: Customer
      required:
        - id
        - name
        - defaultValue
    TemplateWithVersionResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the template
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Template name
          maxLength: 50
          example: Customer Email
        tags:
          description: Array of tag labels
          example: *ref_1
          type: array
          items:
            type: array
        currentVersion:
          type: number
          description: Version number of the latest version
          minimum: 1
          example: 1
        createdAt:
          type: string
          description: Timestamp when template was created (ISO 8601)
          format: date-time
          example: '2026-05-07T10:30:00Z'
        content:
          type: string
          description: Template content with variable placeholders
          example: Hello {{name}}, thank you for your business!
        variables:
          description: List of variables defined in the current version
          type: array
          items:
            $ref: '#/components/schemas/VariableResponseDto'
      required:
        - id
        - name
        - tags
        - currentVersion
        - createdAt
        - content
        - variables
    UpdateTemplateDto:
      type: object
      properties:
        content:
          type: string
          description: New template content (creates new version if provided)
          example: Hey {{name}}, thanks for sticking with us!
        tags:
          description: Updated tag labels
          example:
            - email
            - customer
            - retention
          minItems: 1
          type: array
          items:
            type: array
    TemplateVersionResponseDto:
      type: object
      properties:
        versionNumber:
          type: number
          description: Version number within the template (starting from 1)
          minimum: 1
          example: 2
        templateId:
          type: string
          description: Reference to the parent template
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        content:
          type: string
          description: Template content at this version
          example: Hello {{name}}, thank you for being awesome!
        variables:
          description: List of variables defined in this version
          type: array
          items:
            $ref: '#/components/schemas/VariableResponseDto'
        tags:
          description: Tags assigned in this version
          example:
            - email
            - customer
          type: array
          items:
            type: array
        createdAt:
          type: string
          description: Timestamp when version was created (ISO 8601)
          format: date-time
          example: '2026-05-07T11:00:00Z'
      required:
        - versionNumber
        - templateId
        - content
        - variables
        - tags
        - createdAt
    GeneratePromptDto:
      type: object
      properties:
        templateId:
          type: string
          description: Reference to the template to use
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        variables:
          type: object
          description: Map of variable names to values for substitution
          example:
            name: John Doe
          additionalProperties:
            type: string
      required:
        - templateId
    PromptResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the generated prompt
          format: uuid
          example: 770e8400-e29b-41d4-a716-446655440002
        templateId:
          type: string
          description: Reference to the template used
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        versionUsed:
          type: number
          description: Version number of the template used to generate this prompt
          minimum: 1
          example: 2
        content:
          type: string
          description: Template content with all variables substituted
          example: Hello John Doe, thank you for being awesome!
        variables:
          type: object
          description: Map of variable names to their resolved values
          example:
            name: John Doe
          additionalProperties:
            type: string
        createdAt:
          type: string
          description: Timestamp when prompt was generated (ISO 8601)
          format: date-time
          example: '2026-05-07T14:45:00Z'
      required:
        - id
        - templateId
        - versionUsed
        - content
        - variables
        - createdAt